requirements.txt
streamlit

pandas

gspread==6.1.2

google-auth>=2.0.0

requests